move user-facing formatters to fmt dir
