+++ main/src/ca/uwaterloo/flix/api/lsp/provider/HoverProvider.scala
@@ -19,6 +19,7 @@ import ca.uwaterloo.flix.api.lsp.{Entity, Index, MarkupContent, MarkupKind, Posi
 import ca.uwaterloo.flix.language.ast.TypedAst.{Expression, Root}
 import ca.uwaterloo.flix.language.ast.{Ast, SourceLocation, Symbol, Type, TypeConstructor}
 import ca.uwaterloo.flix.language.dbg._
+import ca.uwaterloo.flix.language.fmt.{Audience, FormatDoc, FormatKind, FormatSignature, FormatType}
 import org.json4s.JsonAST.JObject
 import org.json4s.JsonDSL._
