+++ main/src/ca/uwaterloo/flix/api/lsp/provider/SymbolProvider.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.api.lsp.provider
 import ca.uwaterloo.flix.api.lsp.{DocumentSymbol, Location, Range, SymbolInformation, SymbolKind}
 import ca.uwaterloo.flix.language.ast.TypedAst
 import ca.uwaterloo.flix.language.ast.TypedAst.Root
-import ca.uwaterloo.flix.language.dbg.FormatKind.formatKind
+import ca.uwaterloo.flix.language.fmt.FormatKind.formatKind
 
 object SymbolProvider {
