+++ main/src/ca/uwaterloo/flix/language/ast/Type.scala
@@ -18,7 +18,8 @@ package ca.uwaterloo.flix.language.ast
 
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.Ast.{EliminatedBy, IntroducedBy}
-import ca.uwaterloo.flix.language.dbg.{Audience, FormatType}
+import ca.uwaterloo.flix.language.dbg.FormatType
+import ca.uwaterloo.flix.language.fmt.FormatType
 import ca.uwaterloo.flix.language.phase.Kinder
 import ca.uwaterloo.flix.util.InternalCompilerException
