--- main/src/ca/uwaterloo/flix/language/dbg/Audience.scala
@@ -1,30 +1,0 @@
-/*
- * Copyright 2020 Matthew Lutze
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ca.uwaterloo.flix.language.dbg
-
-/**
-  * Describes the intended audience of the result of a method call.
-  *
-  * [[Audience.External]] indicates the result should be formatted for display to users.
-  *
-  * [[Audience.Internal]] indicates the result should be formatted for display to compiler programmers.
-  */
-sealed trait Audience
-object Audience {
-  case object External extends Audience
-  case object Internal extends Audience
-}
