+++ main/src/ca/uwaterloo/flix/language/errors/KindError.scala
@@ -17,7 +17,7 @@ package ca.uwaterloo.flix.language.errors
 
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast.{Kind, SourceLocation}
-import ca.uwaterloo.flix.language.dbg.FormatKind.formatKind
+import ca.uwaterloo.flix.language.fmt.FormatKind.formatKind
 import ca.uwaterloo.flix.util.Formatter
 
 /**
