+++ main/src/ca/uwaterloo/flix/language/errors/ReificationError.scala
@@ -18,7 +18,8 @@ package ca.uwaterloo.flix.language.errors
 
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast.{SourceLocation, Type}
-import ca.uwaterloo.flix.language.dbg.{Audience, FormatType}
+import ca.uwaterloo.flix.language.dbg.FormatType
+import ca.uwaterloo.flix.language.fmt.{Audience, FormatType}
 import ca.uwaterloo.flix.util.Formatter
 
 /**
