+++ main/src/ca/uwaterloo/flix/language/fmt/MarkDown.scala
@@ -1,0 +1,13 @@
+package ca.uwaterloo.flix.language.fmt
+
+object MarkDown {
+
+  /**
+    * Escapes certain markdown characters.
+    */
+  def escape(s: String): String =
+    s
+      .replace("[", "\\[")
+      .replace("]", "\\]")
+
+}
